Build all packages with Lerna, when building plugin assets
We already have Lerna installed to run linting and tests on all monorepo packages.
Lerna has the option to run commands in the packages in a topological order, this
means it'll follow the dependency tree from the bottom upwards to the start. So if
package A depends on package B, Lerna will first build B before it builds A. This
ensures that the build for package A uses the built version of package B.

This will build both `@yoast/ui-library` and `@yoast/schema-blocks`, so removes the
need for separate build commands for these packages. It also builds `yoastseo` and
`@yoast/analysis-report`. For those packages however, the `main` configuration in the
`package.json` file does not point to the build directory. So for these packages,
the source files will remain to be used.